Resolved conflicts that occurred after the merge

+++ build-front.sh
@@ -10,6 +10,7 @@ fi
 
 # Build the Angular project (using production configuration)
 echo "Building Angular project..."
+npm install
 ng build --configuration production
 
 # Check if build was successful
@@ -25,7 +26,11 @@ cd ../..
 if [ -d "SpringBootServer/src/main/resources/static" ]; then
   echo "Cleaning static directory..."
   rm -rf SpringBootServer/src/main/resources/static/assets
-  rm SpringBootServer/src/main/resources/static/*
+  
+  # Check if static is empty
+  if [ "$(ls -A SpringBootServer/src/main/resources/static)" ]; then
+    rm SpringBootServer/src/main/resources/static/*
+  fi
 else
   echo "Directory SpringBootServer/src/main/resources/static does not exist!"
   exit 1
@@ -35,21 +40,4 @@ fi
 echo "Copying files to Spring Boot static directory..."
 mv WebView/site/dist/site/* SpringBootServer/src/main/resources/static/
 
-# Get the ip
-ip=$(hostname -I | awk '{print $1}')
-SERVER_PATH="http://$ip:8080"
-
-# Path of the JSON file
-JSON_FILE="SpringBootServer/src/main/resources/static/assets/app.config.json"
-
-# Check if JSON exist
-if [[ -f $JSON_FILE ]]; then
-    # Update SERVER_PATH
-    jq --arg newPath "$SERVER_PATH" '.SERVER_PATH = $newPath' "$JSON_FILE" > temp.json && mv temp.json "$JSON_FILE"
-    echo "Le fichier JSON a été mis à jour avec SERVER_PATH = $SERVER_PATH :"
-    cat "$JSON_FILE"
-else
-    echo "Le fichier $JSON_FILE n'existe pas."
-fi
-
-echo "Deployment completed successfully!"
+echo "Frontend building completed successfully!"
